New hardware class
